refactor: Apply code review fixes for solution all-button feature
- Rename solution_category_grouper → solution_category_group
- Fix workbook_url_params.ts: move private helpers after exports, rename param, translate comment to English
- Refactor workbook_url_params.test.ts: split describe blocks by tab type
- Update SolutionWorkBookList.svelte: add category title for single-category view
- Fix E2E tests for updated behavior
