+++ src/features/workbooks/utils/solution_category_group.ts
